# swagger/swagger.json
swagger: "2.0"
info:
  description: "API Rest desarrollada en node.js - Modulo estudiantes"
  version: "1.0.0"
  title: "API REST ESTUDIANTES"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "godaycn87@outlook.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "estudiante-app.herokuapp.com"
schemes:
  - "http"
  - "https"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
basePath: "/"
...
paths:
...
  /inscribirEstudianteExamen:
    post:
      tags:
      - "examenes"
      summary: "inscribirEstudianteExamen"
      description: ""
      operationId: "inscribirEstudianteExamen"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "notas"
        in: "body"
        description: "Parametros listado para inscribirEstudianteExamen"
        required: true
        schema:
          $ref: '#/definitions/inscripcionExamen'
      responses:
        "200":
          description: "OK"
        "406":
          description: "No se encontraron alumnos"
        "404":
          description: "NO EXISTE ALUMNO"
        "500":
          description: "Error al insertar"
      security:
        - Bearer: []                            
  /traerExamenesParaInscripcion/{idEstudiante}:
    get:
      tags:
      - "examenes"
      summary: "traerExamenesParaInscripcion"
      description: ""
      operationId: "traerExamenesParaInscripcion"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "idEstudiante"
        in: "path"
        description: "ID de estudiante"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "OK"
        "400":
          description: "El idEstudiante no puede estar vacio"
        "500":
          description: "ERROR"
      security:
        - Bearer: []    
  /bajaInscripcionExamen/{idInscriptosExamen}:
    delete:
      tags:
      - "examenes"
      summary: "bajaInscripcionExamen"
      description: ""
      operationId: "bajaInscripcionExamen"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "idInscriptosExamen"
        in: "path"
        description: "ID de estudiante"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "OK"
        "400":
          description: "El body no puede estar vacio"
        "406":
          description: "INSCRIPCION FUERA DE FECHA PARA ELIMINAR O NO EXISTE"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []            
  /traerInscripcionesEstudianteExamen/{idEstudiante}:
    get:
      tags:
      - "examenes"
      summary: "traerInscripcionesEstudianteExamen"
      description: ""
      operationId: "traerInscripcionesEstudianteExamen"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "idEstudiante"
        in: "path"
        description: "ID de estudiante"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "OK"
        "400":
          description: "El idEstudiante no puede estar vacio"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []            
  /enviarNotificacionExamen:
    get:
      tags:
      - "examenes"
      summary: "enviarNotificacionExamen"
      description: ""
      operationId: "enviarNotificacionExamen"
      consumes:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "aPartir"
        in: "body"
        description: "Fecha a partir de donde enviar mails"
        required: true
        schema:
          type: "string"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "OK"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []          
...
definitions:
...
  inscripcionExamen:
    type: "object"
    properties:
      idEstudiante:
        type: "integer"
        format: "int64"
      idExamenes:
        type: "integer"
        format: "int64"
      recordatorio:
        type: "integer"
        format: "int64"
    xml:
      name: "inscripcionExamen"  
...